Update v1.6
Upgrading the existing plugin, adapting it to newer versions of the image. The plugin should work on most images. Do not use this plugin for openpli, there is another version just for openpli. Customizing the plugin for the egami image plugin still reports an error on the update feed, but the installation should be successful.